requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cx-Oracle==8.1.0
dnspython==2.0.0
email-validator==1.1.2
fastapi==0.62.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.9.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
orjson==3.4.6
passlib==1.7.4
promise==2.3
pydantic==1.7.3
PyJWT==1.7.1
python-multipart==0.0.5
PyYAML==5.3.1
requests==2.25.0
Rx==1.6.1
six==1.15.0
SQLAlchemy==1.3.20
starlette==0.13.6
ujson==3.2.0
urllib3==1.26.2
uvicorn==0.11.8
websockets==8.1
